libgimpwidgets: Fix stack buffer overrun
3833ab2b1d5055193746a4cb2b4ef864ff15393b changed the type of "rgb" from
gdouble[3] to gfloat[3], but it should instead have changed the type of
"lch".

Resolves #11898.

+++ libgimpwidgets/gimpcolorselect.c
@@ -971,8 +971,8 @@ gimp_color_select_update_pos (GimpColorSelect *select)
 {
   GimpColorSelector *selector = GIMP_COLOR_SELECTOR (select);
   GeglColor         *color    = gimp_color_selector_get_color (selector);
-  gfloat             rgb[3];
-  gdouble            lch[3];
+  gdouble            rgb[3];
+  gfloat             lch[3];
   gfloat             hsv[3];
 
   gegl_color_get_pixel (color, babl_format_with_space ("R'G'B' double", select->format), rgb);
